helper: Don't delete disabled failed jobs

+++ master/extra/delete-failed-jobs.groovy
@@ -6,7 +6,7 @@ pkgJobs = allItems.findAll{job -> job.name.contains("pkg+")}
 for (job in pkgJobs) {
 	build = job.getLastBuild()
 	if (build != null) {
-		if (build.result != Result.SUCCESS) {
+		if ((job.isBuildable() == true) && (build.result != Result.SUCCESS)) {
 			println("Delete: ${job.getName()}");
 			job.delete();
 		}
